# [Sanity Check] Run query once per search