Not urgent, but if I'm on the right track, we can improve search performance by a fair bit.

---

It looks like controllers/catalog/Search.php is running each search query twice: once for the items on the current page, and then again to count the total number of matches for the search.  There doesn't seem to be a clean way to count the total results, but here's _one_ way.

We can't use a regular `COUNT(*)`, because that will stop at our `LIMIT` of 25.  So: the `OVER()` call ([1], [2]) lets us `COUNT` before we discard the other results.
But: doing this on each of our `UNION` sets will give different results, which we'd need to add together selectively.  Collecting them all in a sub-query (`WITH results AS (`) gives us a combined set to count.

I don't know how much of a hack this is, and how much overhead it adds, but it at least seems much faster than running the entire search query twice.

[1]: https://stackoverflow.com/a/8242764
[2]: https://stackoverflow.com/a/28888696