@notartom 2024-03-16
I'm sure there are many other things wrong with search that we can improve and fix and change, but as small incremental improvements go, this one is pretty good. @redrun45 you wrote [Sanity check], are you opposed to merging this?