@redrun45 2024-03-16
@notartom if the technical direction is sound, then no, I'm not opposed.  I just have to patch the simplesearch part too.  I'll get a second commit ready soon!